Move "build" to daemon/build.go
